# semester1/week4/ginserver1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "这是一个基于Gin框架的用户管理系统API，提供用户注册、登录、个人信息管理等功能",
    "title": "用户管理API",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded"
  ],
  "produces": [
    "application/json",
    "text/html"
  ],
  "paths": {
    "/changepassword": {
      "get": {
        "description": "返回修改密码页面",
        "produces": [
          "text/html"
        ],
        "tags": [
          "页面"
        ],
        "operationId": "changePasswordPage",
        "responses": {
          "200": {
            "description": "成功返回修改密码页面"
          }
        }
      },
      "post": {
        "description": "修改用户密码",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "text/html",
          "application/json"
        ],
        "tags": [
          "用户"
        ],
        "operationId": "changePassword",
        "parameters": [
          {
            "type": "string",
            "description": "用户名",
            "name": "username",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "原密码",
            "name": "oldpassword",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "新密码",
            "name": "password",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "确认新密码",
            "name": "password1",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "密码修改成功",
            "schema": {
              "$ref": "#/definitions/APIResponse"
            }
          },
          "400": {
            "description": "用户不存在或两次密码不一致或信息为空",
            "schema": {
              "$ref": "#/definitions/APIResponse"
            }
          },
          "403": {
            "description": "原密码错误",
            "schema": {
              "$ref": "#/definitions/APIResponse"
            }
          },
          "500": {
            "description": "服务器内部错误",
            "schema": {
              "$ref": "#/definitions/APIResponse"
            }
          }
        }
      }
    },
    "/changeprofiles": {
      "get": {
        "description": "返回修改个人信息页面",
        "produces": [
          "text/html"
        ],
        "tags": [
          "页面"
        ],
        "operationId": "changeProfilePage",
        "parameters": [
          {
            "type": "string",
            "description": "JWT认证令牌",
            "name": "token",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "用户名",
            "name": "username",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回修改个人信息页面"
          },
          "400": {
            "description": "认证令牌和用户名不能为空",
            "schema": {
              "$ref": "#/definitions/APIResponse"
            }
          },
          "401": {
            "description": "无效或已过期的认证令牌",
            "schema": {
              "$ref": "#/definitions/APIResponse"
            }
          }
        }
      },
      "post": {
        "description": "修改用户个人信息",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户"
        ],
        "operationId": "changeProfile",
        "parameters": [
          {
            "type": "string",
            "description": "原用户名",
            "name": "username",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "新用户名",
            "name": "newusername",
            "in": "formData"
          },
          {
            "type": "string",
            "description": "新姓名",
            "name": "newname",
            "in": "formData"
          }
        ],
        "responses": {
          "200": {
            "description": "信息修改成功",
            "schema": {
              "$ref": "#/definitions/APIResponse"
            }
          },
          "400": {
            "description": "未发生更改或用户名/姓名与原值相同或信息为空",
            "schema": {
              "$ref": "#/definitions/APIResponse"
            }
          },
          "500": {
            "description": "服务器内部错误",
            "schema": {
              "$ref": "#/definitions/APIResponse"
            }
          }
        }
      }
    },
...
    "/login": {
      "get": {
        "description": "返回用户登录页面",
        "produces": [
          "text/html"
        ],
        "tags": [
          "页面"
        ],
        "operationId": "loginPage",
        "responses": {
          "200": {
            "description": "成功返回登录页面"
          }
        }
      },
      "post": {
        "description": "用户登录并返回JWT令牌",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户"
        ],
        "operationId": "loginUser",
        "parameters": [
          {
            "type": "string",
            "description": "用户名",
            "name": "username",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "密码",
            "name": "password",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "302": {
            "description": "登录成功，重定向到个人资料页面"
          },
          "400": {
            "description": "用户不存在或信息为空",
            "schema": {
              "$ref": "#/definitions/APIResponse"
            }
          },
          "403": {
            "description": "密码错误",
            "schema": {
              "$ref": "#/definitions/APIResponse"
            }
          },
          "500": {
            "description": "服务器内部错误",
            "schema": {
              "$ref": "#/definitions/APIResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "APIResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        },
        "data": {
          "type": "object"
        }
      }
    }
  },
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT认证令牌，格式为\"Bearer {token}\""
    }
  }
}